+++ services/director-v2/src/simcore_service_director_v2/modules/osparc_variables/substitutions.py
@@ -4,7 +4,6 @@ import functools
 import logging
 from copy import deepcopy
 from typing import Any, Final, TypeVar
-from urllib.parse import urlparse, urlunparse
 
 from fastapi import FastAPI
 from models_library.osparc_variable_identifier import (
@@ -24,7 +23,9 @@ from models_library.wallets import WalletID
 from pydantic import BaseModel
 from servicelib.fastapi.app_state import SingletonInAppStateMixin
 from servicelib.logging_utils import log_context
-from servicelib.rabbitmq.rpc_interfaces.webserver.products import get_product_base_url
+from servicelib.rabbitmq.rpc_interfaces.webserver.products import (
+    get_product_api_base_url,
+)
 
 from ...modules.rabbitmq import get_rabbitmq_rpc_client
 from ...utils.db import get_repository
@@ -40,16 +41,6 @@ from ._user import request_user_email, request_user_role
 _logger = logging.getLogger(__name__)
 
 TBaseModel = TypeVar("TBaseModel", bound=BaseModel)
-
-
-def _make_api_server_base_url(base_url: str):
-    parsed = urlparse(base_url)
-    hostname = parsed.hostname or ""
-
-    if not hostname.startswith("api."):
-        hostname = f"api.{hostname}"
-
-    return urlunparse((parsed.scheme, hostname, "", "", "", ""))
 
 
 async def substitute_vendor_secrets_in_model(
@@ -217,11 +208,9 @@ async def resolve_and_substitute_session_variables_in_model(
                 node_id=node_id,
                 run_id=service_run_id,
                 wallet_id=wallet_id,
-                api_server_base_url=_make_api_server_base_url(
-                    await get_product_base_url(
-                        get_rabbitmq_rpc_client(app),
-                        product_name=product_name,
-                    ),
+                api_server_base_url=await get_product_api_base_url(
+                    get_rabbitmq_rpc_client(app),
+                    product_name=product_name,
                 ),
             ),
         )
@@ -267,11 +256,9 @@ async def resolve_and_substitute_session_variables_in_specs(
                     node_id=node_id,
                     run_id=service_run_id,
                     wallet_id=wallet_id,
-                    api_server_base_url=_make_api_server_base_url(
-                        await get_product_base_url(
-                            get_rabbitmq_rpc_client(app),
-                            product_name=product_name,
-                        ),
+                    api_server_base_url=await get_product_api_base_url(
+                        get_rabbitmq_rpc_client(app),
+                        product_name=product_name,
                     ),
                 ),
             )
